Update thrift version to 0.10.0
Thrift version 0.9.2 includes an unnecessary production
dependency on 'nodeunit', which in turn pulls in dependencies
on 'tap' and 'nyc'.

In addition to the obvious problems, this can cause issues for
users of 'yarn', due to poor interactions with 'yarn check' and
'nyc's use of 'bundledDependencies'.

This uses version 0.10.0 of thrift rather than the more recent
0.11.0 release, due to the latter's use of a version of 'ws'
not suported in older node versions.