allow specifying which auto-instrumented urls should have tracing headers (#237)
* allow specifying which auto-instrumented urls should have tracing headers

While the tracing headers are useful when sent to backends that expect, and
utilize them, automatic instrumentation is also useful for just tracking the
time and response status of requests to external services that aren't being
tracked. If such an external service has a strict policy on allowed headers the
automatically added headers will cause instrumented requests to error. So
instead of not getting any information from those URLs this commit adds the
ability to whitelist, and blacklist, the URLs that have tracing headers added to
them.

The new options are built off of the existing blacklist options and are:

- fetch_url_header_inclusion_patterns
- fetch_url_header_exclusion_patterns
- xhr_url_header_inclusion_patterns
- xhr_url_header_exclusion_patterns

* refactor _shouldTrace methods